prometheus: set v2 max block duration to 24h
Defaults to 10% of retention time, however we're usually running instances with
hundreds of days of retention. For a block to be purged from storage its time
span needs to be fully outside retention period.

Fix 24h as max block duration so that expiration logic can eliminate blocks on
an ~daily basis.

Bug: T187987
Change-Id: I736de08e3d46cdf1f3c76ff20ff42b44b1fa9759