mariadb-backups: Allow remote dumps from cumin hosts
dumps used to be only initiated from the host where the backups
were stored. Enable dumps also from cumin hosts, using the same
workflow than snapshots, but without the previous transfer.

It is yet no unclear if this will be the final workflow, but it
should at the very least, work.

Further work on scheduling several backups on the same
provisioning host, and coordinating dumps and snapshots may be
needed, but for now they are all treated equally.

To test remote dumps, one has been scheduled to run on dbprov2001
to test it, not intended to be the final configuration.

Bug: T206203
Change-Id: Ia074122af3e64e0b93665ed959dd17b357950652