openstack: serverpackages: mitaka: stretch: ignore python-cryptography from bpo
Ignore the python-cryptography package that comes from jessie-backports (as
imported into jessie-wikimedia/openstack-mitaka-jessie).

That version creates a conflict with a previous rule we had to don't install
libssl1.0.0 in stretch systems. By the time of this patch, I don't remember
why is there an explicit rule to uninstall libssl1.0.0 from jessie-backports,
but is likely that it was related to a security upgrade.

=== 8< ===
[ in a cloudvirt stretch system ]
$ aptitude why python-cryptography
i   python-nova Depends python-cryptography (>= 1.0)

$ apt-cache policy python-cryptography
[..]
  Candidate: 1.7.1-3~bpo8+1
  Version table:
     2.3-1~bpo9+1 100
        100 http://mirrors.wikimedia.org/debian stretch-backports/main amd64 Packages
 *** 1.7.1-3 500
        500 http://mirrors.wikimedia.org/debian stretch/main amd64 Packages
        100 /var/lib/dpkg/status
     1.7.1-3~bpo8+1 1001
       1001 http://apt.wikimedia.org/wikimedia jessie-wikimedia/openstack-mitaka-jessie amd64 [..]

$ aptitude show python-cryptography=1.7.1-3 | grep Depends
Depends: [...] libssl1.1 (>= 1.1.0)
$ aptitude show python-cryptography=1.7.1-3~bpo8+1 | grep Depends
Depends: [...] libssl1.0.0 (>= 1.0.1)
=== 8< ===

So, we are interested in the native stretch version, not in the version
included in jessie-wikimedia/openstack-mitaka-jessie.

An obvious alternative to this patch is to drop python-cryptography from the
jessie-wikimedia/openstack-mitaka-jessie repo, but we can't because it's used
in jessie systems:

=== 8< ===
[ in a cloudvirt jessie system ]
$ aptitude why python-cryptography
i   nova-common Depends python-nova (= 2:13.1.0-2~bpo8+1)
i A python-nova Depends python-cryptography (>= 1.0)

$ apt-cache policy python-cryptography
python-cryptography:
  Installed: 1.7.1-3~bpo8+1
  Candidate: 1.7.1-3~bpo8+1
  Version table:
 *** 1.7.1-3~bpo8+1 0
       1001 http://apt.wikimedia.org/wikimedia/ jessie-wikimedia/openstack-mitaka-jessie amd64 [..]
        100 /var/lib/dpkg/status
     0.6.1-1+deb8u1 0
        500 http://mirrors.wikimedia.org/debian/ jessie/main amd64 Packages
=== 8< ===

Hosts: cloudvirt1014.eqiad.wmnet, cloudcontrol1004.wikimedia.org, cloudnet1003.eqiad.wmnet

Change-Id: I9bb6cad7c60c2cf0a4f4a60ba108c93705f1d2af
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>