wmcs: introduce codfw1dev puppet code
Introduce roles/profiles/hiera for this new deployment. The deployment is not
new actually, we are just renaming/rebuilding it from the old labtestn.

Most of the hiera keys will need a later update as wel start fixing stuff, but
the current values can serve as placeholders meanwhile.

More info here:
 https://wikitech.wikimedia.org/wiki/Portal:Cloud_VPS/Admin/Deployments

Bug: T219626
Change-Id: If286b77950479c8589d9635a68e5ef60c856f691
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>