compiler-update-facts: better support addition of arbitrary fact sets
Previously this script only supported aggregating facts from a single
puppetmaster (on subsequent runs it would clean up any pre-existing facts
from other masters).

This change leaves previous behavior as it was, but also allows the
user to pass in an arbitrary puppetmaster (e.g. a project-local WMCS
master); in that case facts are uploaded into a subdir that is left
unaltered by updates from other puppetmasters.

This should allow us to maintain fact lists for different puppetmasters
without any one user of this script needing to be aware of/have access
to masters outside of their interest.

Note that once this script is applied there will need to be one manual
purge of the facts in the top-level yaml dir on the compiler workers.

Bug: T219430
Depends-On: Ib980edec419a30dde3eb0f3317d2d8136b7d8db7
Change-Id: I144fef9c93223b780cea2a83ade2e7b2a51ce964