pdns-recursor: reduce maximum number of file descriptors
On a host running an older version of the recursor (3.5.3) I'm
seeing some very bad system-wide behavior following a surge in
dns requests.

We already reduced this number to 512 on the advice of the linked
docs, but that still lets us hit 1024 file descriptors.  Given
that 1024 is probably the per-process limit, it seems like
pushing our luck to hit 1024 on the nose.  With this change
we're limited to u sing 768 threads for handling requests,
which should leave plenty of headroom in case the recursor
wants to do even one additional thing.

Note that this change is surrounded by a guard for version
4 or earlier, which is to say: only on Trusty in our world.

Bug: T219953
Change-Id: Icd31ea17785c3d160668b5817c63bdf92170981e