openstack: add some missing bits for codfw1dev
In case a stretch-based openstack deployment is using a local database.
This is the case of codfw1dev. Also, some hiera keys were missing (or on
the contrary, they don't make sense here).

While at it, drop support code for shared deployments, since we won't have
shared deployments in the short term in codfw.

Bug: T219626
Change-Id: If1fc640b92fcedfbea32818edbc1a47088e23286
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>