pdebuild: ensure proxy config exists for apt-get install
This CR ensures the proxy configuration for archive.debian.org is in
place for apt get install commands.  Acquire::Check-Valid-Until does not
support per domain configuration.  As we only need this for the update
command and we dont want it to be global we dont write it to a config
file

Change-Id: I835260b7efbbe18c1474ff3858705c128cdd03ac