mariadb-backups: Setup dbprov2002
dbprov2002 is the new database provisioning, short-term storage
and backup-generation host; this is in addition to the recently
setup dbprov2001.

Because now there are 2 provisioning hosts, we have to coordinate
both. For now snapshots will just be coordinated in a single
process (optimizations could be done later so it spawns an
independent process for each target host).

For dumps, we decide what configuration will go on each host, for
that we create a configuration file per host, rather than per
dc. This is ok because if a new host is configured, puppet will
complain due to the lack of the template. As there is still only
1 host on eqiad, that template only gets renamed, with no change.

Once this is deployed, backups will be divided in the following
way:

dbprov2001 dbprov2002
s4         s1
s5         s2
s6         s3
s8         s7
x1         m1
m2         m3
m5         tendril

Mostly decided based on size and not combining problematic ones
on the same host (s1, s4, s8), (x1, s3), m2 size, etc

Bug: T218336
Change-Id: Ica50a8cd7a854217852d4f060c7f7e0aad5eb372