uwsgi: allow setting routing rules
As the UwsgiHandler in apache is very raw, we need to be able to use the
routing logic UWSGI allows us to have in order to set things like
e.g. HTTP response headers.

Given routes can and will be repeated in the configuration, allow to
provide them via an array rather than a hash.

Change-Id: I7bfb2e0b870737caf4399722f4dd6705b80bbfb4