Add qemu processes/Ganeti instances to filter_services list of debdeploy
Each of these processes represents a Ganeti instance, but restarting those
is outside of the restarting handling of debdeploy; they either need to
be restarted via "gnt-instance reboot" or powered down/powered up. If
we have a specific reason to orchestrate such instance restarts, it's
coordinated, but in the general case we can filter them from the restarts
list.

Bug: T135991
Change-Id: I84a1587f481bbd79b0f8e6798e3f1046718ffba2