cloudstore: start refactor for role switch up around the labstores
It turns out that the reason the primary cluster is called the secondary
cluster is that the primary cluster got turned into spares and things were
never updated.  This is just changing the naming of the class and all the steps
needed to accomplish that.
The point here is that after this is changed and cleaned up, I can structure
the roles and the labstore module so that the new "secondary" cluster can
reuse most of the labstore module--maybe even correct all the hiera todos.
One nice thing is that will allow nfs-exportd to work for the maps servers
(which would admittedly been more useful during trusty deprecation, but will still
be nice during stretch deprecation down the road).

Bug: T209527
Change-Id: Icc89332f0e779f6c924119c46edd922e9e698aa9