mariadb: Move wikimedia_editor_tasks_entity_description_exists
wikimedia_editor_tasks_entity_description_exists no longer lives on x1,
but on local database. And it is not private.

Move it to filtered_tables.txt for now although there will be no views
for now

This requires to restart sanitariums and copy the table from the
sanitarium master to sanitarium (with replication so it arrives to labs)

Bug: T218302
Change-Id: I726f7f1ea95af056278f82b9b5894e0abc3f01af