openstack: clientpackages: fix missing deb repo installation
This patch affects 2 very different use cases:
 * openstack clients (all Cloud VPS instances, and beyond)
 * openstack HW servers (i.e, cloudcontrol, cloudnet, cloudvirt, etc)

The issue this is fixing is related to the openstack HW servers.

The clientpackages profile is evaluated first for new HW servers. This is
problematic because the openstack::clientpackages::common class doesn't include
any specification of the openstack version, just blindly installs some client
libs. If concrete openstack repo/pinning configuration is not present, packages
are installed from an arbitrary repo, from whatever openstack version is
present in that candidate repo.

In the case of HW servers, later on in the first puppet run, the proper
repo/pinning configuration for this openstack/debian version is installed, but
at this point puppet agent won't downgrade packages already installed. This
results in a complete mess in the packages being installed in the servers, a
mixture of packages from different openstack versions and repos.

Part of why this was missed before is the overloading of the 'common' word,
so I'm renaming the affected class to a more meaningfull name:

* from -> openstack::clientpackages::common
* to   -> openstack::clientpackages::anyopenstack_anydebian

Usually openstack clients packages and libs aren't very strict about
versioning, so this change should have little effect for openstack clients
but is fixing a dissaster in HW servers, which must run a concrete openstack
version.

For consistency, this new class is required in all
openstack::clientpackages::${openstack}::${debian} subclasses, along with the
class that should include the repo/pinning configuration (for mitaka only by
now).

This patch also introduces openstack::commonpackages::ocata because there
are some servers running ocata already (horizon?).

Change-Id: I9f6524499c817c1439f868d0e32c9a408f650862
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>