openstack: designate: introduce openstack/debian split layout
This allows us to reuse the serverpackages code from here, instead of including
the profile in the role.

This patch adds other benefits:
 * we drop -t jessie-backports, which no longer makes sense.
 * we add initial support for stretch (not tested with this patch)

A follow-up patch should make sure the split is completed, since this patch
is part of a unrelated factorization and I'm focusing right now in producing
a minimal puppet catalog diff with it.

Change-Id: I6ad377c472621dbe5a98775752efd0088d25691f
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>