openstack: drop serverpackages profile
Not really used, since all configuration and package installation is
done from concrete server modules (keystone service, nova service, etc), which
in turn include the base openstack::serverpackages::XXXX (which is where the
big deal is, instead of the profile).

The purpose of this profile was to allow per-deployment overrides to the
openstack server packages. But we don't need them now and we can re-introduce
them later if we really require them.

Having this code hanging around just creates confusion, can lead to mistakes
(such as duplicate resources), etc.

Change-Id: I674545101b9914343003249ce7d6ba61a9ae4d43
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>