openstack: codfw1dev: don't include clientpackages in cloudcontrol servers
Most client packages are installed by DEB dependencies of openstack server
components anyway.

The clientpackages profile is evaluated first for new HW servers. This is
problematic because the openstack::clientpackages::common class doesn't include
any specification of the openstack version, just blindly installs some client
libs. If concrete openstack repo/pinning configuration is not present, packages
are installed from an arbitrary repo, from whatever openstack version is
present in that candidate repo.

In the case of HW servers, later on in the first puppet run, the proper
repo/pinning configuration for this openstack/debian version is installed, but
at this point puppet agent won't downgrade packages already installed. This
results in a complete mess in the packages being installed in the servers, a
mixture of packages from different openstack versions and repos.

This patch drops the included/required profile only in the codfw1dev
deployment. Probably eqiad1 will follow.

Bug: T219981
Change-Id: I0eabcb5f39b94abc411ee912c2444a22300ca13f
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>