admin: temporary remove piccardi from analytics-privatedata-users
User 'piccardi' sent for two times a big and problematic job to the hadoop
cluster that tore down the HDFS masters. We tried to contact him but he seems
to be unreachable via the email listed among his contacts, removing SSH
access to analytics nodes as protective measure until we can contact him.

Change-Id: Iac672f3e94acb973e2a60bacb29a4e9b0d87117d