postgresql: set recovery.conf to writeable by postgres user
When doing a failover, I followed common practice and ran pg_ctl promote
as the postgres user.  This resulted in the database crashing because
it could not rename recovery.conf to recovery.done.  While it may be
a documented boobytrap, it is, indeed a boobytrap on failover because it
crashes the postgresql service until permissions are changed and this is
re-attempted (or possibly it is run as root directly).

Unless there is a very strong reason not to change this (and there might be!)
I am proposing we make the file writeable to the owner and set the owner to
postgres.  However, I'm very open to the idea of simply documenting for WMCS
where we'll find it to only do failovers as root!

Bug: T219652
Change-Id: I433182072fb651acf205fd7a87e29879ada437d7