postgresql: set max_wal_senders on slave conf
When a slave is promoted, it will not be able to replicate back to the new
replica until it is restarted again after a puppet run with the current
config.  max_wal_senders cannot be changed except via database restart, which is
disruptive.  Reload won't cut it.

With this setup, the file will be sufficient for smooth failover.

Bug: T219652
Change-Id: I87698d90cfb351cfb76aecf65f71414717173383