admin: remove sudo permissions from gpu-testers and add users to it
The new GPU on stat1005 seems working and we decided to start moving
stat1005 to a more production-ready host (rather than a testing env).

After a chat with Erik we decided that root permissions are not needed
anymore, so removing them from gpu-testers.

Miriam and Gilles are interested in working/testing with the new AMD GPU,
so adding them to the group as well to grant ssh access.

Bug: T148843
Change-Id: I0045fee07f070bba0b7cc98897fbd4c0a5323413