admin: create analytics-deployers group
This group will be used to allow people to hold the analytics_deploy
keyholder's key to deploy via scap to Analytics misc nodes. We are currently
allowing only analytics-admins, but we had recently a use case related
to deploying to stat1007 a performance team's repository to control a specific
public dataset, published via analytics' UIs. Rather than adding Gilles to
a group with a lot of sudo permissions only for deploying, I propose to create
a specific analytics-deployers group to be used instead.

Bug: T220175
Change-Id: I0aa8b380a5b85ee6e17d4c4ab18b8954204b5af3