puppet: Refactor of the base::puppet class
I got a little rabbit holed today starting out with what i thought would
be a small refactor. it turned out to be a bit bigger then that and
may not be consistent with current coding styles.  however i have tried
to adhere to the puppet coding guides.  Comments most welcome

Bug: T219803
Change-Id: Idd0918f9ed204460d01815f338ecfdffc55c8da8