icinga/parsoid: disable systemd monitoring on scandium test host
Currently the parsoid-vd service has a problem on the parsoid
testing host scandium.

We want to remove monitoring of systemd state for now as requested
by subbu. To be possibly re-enabled later.

We do want to keep all the other base monitoring though.

Fortunately this was already done for gerrit2001 in the past so
now it's just a simple Hiera setting to turn this off/on.

Bug: T219933
Change-Id: I1fe4aba52a2e05ec8e2d7d6b17b4dcc652c740d5