zuul: stop pinning python-pbr
The zuul version comes from the Debian package version. It was
compatible not compatible with python-pbr 1.10 and we had to pin it to
the verison that comes from Jessie.

Eventually Zuul 2.5.1-wmf6 gets the version from Distribute instead of
relying on python-pbr.  We thus no more need the package pinning.

Bug: T218559
Change-Id: I21895ddd66f03520682c867f40142427c646fed0