service::node: Allow the world to read the config
chmod'ing to 0440 allows only the deployment user to read the config,
but not the run user as well, which causes services to block once
restarted. This commit chmod's to 0444 to allow the service to read
their own config.

This is currently just a work-around, a more proper solution would be to
add the service users to the deployment group (other ideas welcome too).

This is a follow-up on I665f6a4536dac056cae20d1b0effb0f0f8bd1d3e

Bug: T207143
Change-Id: Ib230dc3d0f4a1dd5307fb6d51e41aa25b1e137ba