striker: factor out common code to a shared profile
Instead of declaring stuff directly in the role, which goes against our puppet
coding guidelines.

Create `profile::wmcs::striker::web` with all the shared code and include it in
previous caller roles.

This refactor should be a PCC noop.

The class role::striker::web is explicitly left out of this factorization since
that is used in the labs realm.

Change-Id: I116ea835a0cac2ded228d6d01a768f31b5f467cf
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>