role::druid::public::worker: set stricter timeouts
We have been observing connections piled up and timing out when
a Historical is under query load and some segments are
scheduled to be deleted. The default values are very high and
probably not right for the kind of traffic that Druid Public
is receiving (namely the ones coming from AQS).

Bug: T219910
Change-Id: Id4ed9da505b785c0890ce54413045036f195a44f