sessionstore: create profile class to fix password lookups
After adding the new variables to Hiera in Ief835cf57222839038f6
his finally fixes the password lookup issue described in the ticket.

If the password include is not in the profile we don't have them
before profile::cassandra does
$cassandra_settings = hiera('profile::cassandra::settings') and
the passwords stay blank.

Including the passwords in the role is against style guide so
we are creating a mostly empty profile class for sessionstore and
get the passwords there. This is also the difference versus existing
code using cassandra; there was a profile class.

Bug: T219560
Change-Id: I99ae749585ba3dd9853e147027f023dc04b13c5c