cache: add profile::cache::varnish::frontend
Begin disentangling the frontend part of cacheproxy::instance_pair by
moving the varnish::instance("${cache_cluster}-frontend") call to a new
profile called cache::varnish::frontend.

Use hiera instead of code to distinguish between text and upload.

We are going to do the same with profile::cache::varnish::backend in the
next commit.

The end goal is to be able to get rid of the profile::cache::text/upload
distinction, and of cacheproxy::instance_pair too.

Bug: T219967
Change-Id: If8cb3b11d496786017b27c6de5d4a0e3e441ea93