openstack: virt: reallocate libssl1.0.0 package exclusion
In cloudcontrol servers we may have a dependency cycle requiring libss1.0.0,
but for whatever reason we exclude it in cloudvirts servers. Move this config
to an openstack component which is only used in cloudvirts so cloudcontrol
servers aren't affected by the config.

For the record, I don't know why libssl1.0.0 is exluded at all. The commit in
which was introduced (by me at 5614cce1b9d13c62cf4d13b6906bea239f1b6520)
doesn't give much details, but I can think on the typical dependency cycles,
or even worse, nova refusing to work with that concrete libssl version.

Bug: T215407
Change-Id: Ic703c2aa1ec18805ee4804aea1028ac7c3afcd33
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>