facter3/puppet5: remove defaults from hiera
Currently hiera is configured such that the nuyaml backend takes
priority over the role3 backend.  As such if a value is specified in any
of the files looked up using the nuyaml backend then the role3 backend
will not get called.  As such it means we cant store default common
values in hiera if we want to override them at the role level.

Change-Id: I0a1e16be413cd3db9ce946dd538adeebdc165c8b