admin: allow analytics-admins to control jupyter user units
The SWAP nodes (notebook100[3,4]) allow people to run their
Jupyter notebooks, created via a UI. From the system point of view,
there is a daemon called jupyterhub that controls all this flow,
and some temporary systemd units with the following nomenclature:

jupyter-$username-singleuser.service

It happens a lot of times that somebody in Analytics is pinged to
debug some issue on other's units, that often requires to fix
some config and restart the systemd unit. This is currently only doable
by opsens, so extending the permissions to whole Analytics team as well.

Change-Id: If8299cb7f111d51e3815f0c428e1f4a9065c00b8