profile::hadoop::spark2: auto upload of spark2-assembly.zip
It seems that the spark2 defaults require a spark2-assebly.zip file
present in Hadoop (to allow Yarn to distributed the necessary Spark2
jars around). We upload it only in the cdh::spark class afaics, that is
now deprecated. Adding a specific step in the Spark2 profile.

Also refactor the location of other files to have a better organization
in the puppet repository.

Bug: T218343
Change-Id: I5c8f33fda92db97e005492d305bd073069ef9ced