profile::kerberos:*: only use global variables
I came up with a patch to configure Kerberos for the hadoop test
cluster, and realized that since we'll likely have only one realm
it would make sense to have a global config in a place like common.yaml
in hiera to avoid unnecessary repetitions. This is what Moritz has been
trying to do in previous code changes so apologies for the change of mind.

Moreover the kerberos::something hiera argument in the client profile
seems not picked up by hiera if set in commons.yaml.

Change-Id: I55f791bf03e381da529a53fe42a4c2df0967d202