gerrit: increase core.packedGitLimit
This is the size of git packfiles to keep in memory at any given
time (as part of the heap). Our top-10 most fetched repositories
packfiles total roughly 2g.

A majority of traffic to gerrit is upload-pack (fetches). To keep those
fetches fast, we should try to keep their packfiles in memory as much as
possible. The median size of the packfiles for our top 10 repos is 36m.
We had 109 repos fetched > hourly yesterday. Back of envelope math: 109
repos * 36m/repo is roughly 4g.

I have seen packedGitLimit recommendations of anywhere from 2g up to 1/2
your heap (10g for us). This feels like a conservative place to start.

Bug: T221026
Change-Id: Ie523c331d4931304f2b096c8fe5dffb3ac61a41c