gerrit: lower sshd threadpool size
sshd.threads in Gerrit, counter intuitively, defines the limit of
concurrent git operations over both https and ssh.

The default is 2 * <number-of-cores> which is 32 for cobalt.

In Gerrit versions before 2.13 this limit was, by default, 1.5 *
<number-of-cores>.

We upgraded to 2.13 at the end of 2017 which implicitly bumped our
sshd.threads from 24 to 32.

This lowers the thread limit to previous levels for interactive users.
This also sets the number of "batchThreads" available to non-interactive
users. This number is subtracted from the sshd.threads count (leaving us
with 24 threads for interactive users) and 4 threads for dedicated use
by non-interactive users (e.g., jenkins-bot and phab).

Bug: T221026
Change-Id: I9251e17c9e6c94ec408c5a3f6e342f7e7b6ed1c6