icinga/base/prometheus: add notes_url to DISK space checks
As requested in T220326#5091992 adding a notes_url
to link to runbooks for prometheus checks.

Since this is a generic check added to everything in
the base module, i made a landing page that is
suppoed to have sections/links to runbooks specific
to a certain type of servers / role.

https://wikitech.wikimedia.org/wiki/Monitoring/Disk_space ->
https://wikitech.wikimedia.org/wiki/Prometheus#Prometheus_host_running_out_of_space

Bug: T220326
Change-Id: I9f1eb0c2e97da38dec2a9794179cef0f7ce9cbf6