puppetdb: adapt the module so it works on Cloud VPS simply again
We'd been testing this in toolsbeta for some time, but since it's a test
env, we didn't notice that puppet was broken for quite a while on the
puppetdb server.  Therefore, this is aiming to make more of the new values
and services optional (such as ELK logging and the microservice for netbox).

The aim is to have it working without a lot of hiera config that doesn't
apply so it can be used in projects.

This should help that whole idea, but the fix for the initial problem in
toolsbeta specifically would be I16d886ede96dd0b

This should be a NOOP on production pupppetdb hosts, or it's broken.

Bug: T221721
Change-Id: I1b5f9326a6e86adcd1854cc799181c186830d27e