wmcs: Respect LDAP locks in ssh-key-ldap-lookup
Configuration for wikitech was recently updated to include LDAP account
locking as part of on-wiki blocks. When a user is blocked site-wide on
Wikitech we now add a custom pwdPolicySubentry attribute to their
associated LDAP account which prevents password based authentication.
Our sshd integration is key based rather than auth-bind based, so to
have the same effect (disable logins) in the Cloud VPS realm we need to
check for this attribute before collecting keys and providing them to
sshd. This change does that by adding a search filter to the LDAP lookup
which will exclude any records having a pwdPolicySubentry attribute with
the value "cn=disabled,ou=ppolicies,dc=wikimedia,dc=org".

Bug: T168692
Bug: T218654
Change-Id: Ifd6eefc789c352b97e0b4ee6573bef4031914734