zuul: log stack dump to their own file
When sending SIGUSR2 to the Zuul scheduler process, it emits a
stacktrace to the logging system. The qualifier is 'zuul.stack_dump' and
the message is emitted at DEBUG level.

Capture that to a /var/log/zuul/stack_dump.log file which is rotated
every 7 days.

Hosts: contint1001.wikimedia.org

Change-Id: Ic38d535c05255c3ee9270af9169842cb30810f86