cloudstore: add python3 clientpackages for all
Because it is strangely impossible to override the openstack version
for clientpackages on stretch, and our scripts by and large require
python3 packages, add the python3 packages in common.

This is somewhat frustrating because the packages that actually get
installed are Newton, not Mitaka, but hiera couldn't be convinced?

I'd set:
profile::openstack::eqiad1::version: newton
profile::openstack::base::version: newton
profile::openstack::main::version: newton
on the role, but it wouldn't take.  Whether I'm doing that wrong or not,
The python3 packages are required for any clientpackages setup to function
right with cloudstore/labstore scripts :)

Bug: T209527
Change-Id: Id9c002440cdb381b8ebb7014f69320a774bcda3c