auth pdns: bind on IPv6
The new cloud-ns servers have IPv6 addresses and the relevant AAAA records, but
we're not listening on them and this is the source of issues. We could just
remove the IPv6 addresses from DNS but it's 2019.

The exact effect this has on pdns is untested, this is based on the docs.

Bug: T221527
Change-Id: Ibd9c2f913129779d42ff6b41ab1893cc449c9535