mariadb: Promote db2079 to s8 codfw master
With all the new hosts arriving to codfw, let's start some work on
leaving codfw as similar as eqiad as possible.

s8 in codfw has all the same hosts as eqiad, so it is ready to have a
new host as master.
db2045 will be eventually decommissioned in 2020 (so it can be used
somewhere else, like x1 where there are some other hosts that need to be
decommissioned soon.

Bug: T220170
Change-Id: Iccca22ccdc994fd0a5427cf99ed3cd077ac940a3