thumbor: Use port 8800 for haproxy
We will be replacing nginx with haproxy to proxy
requests back to thumbor servers. The easiest way to
achieve this would be to have haproxy listening to the
port where nginx was previously listening on.

This will require 2 puppet runs to work properly and it
will cause service disruption. For that reason we will
first disable puppet on all thumbor hosts, and then
depool each one and run puppet.

Bug: T187765
Change-Id: Iecab5f2e839e1013c559e573328a2e50a37f5a04