openstack: clientpackages: split profile for VM instances
Previous to this patch, the clientpackages profile was being used for both VM
instances and hardware servers. This leads to complex interactions and deadlock
situations.
We would rather have to hability to modify the profile independently for VMs
and HW, so split the profile in two.

This patch introduces a new base openstack profile meant to be used only for VM
instances in CloudVPS. Following this approach, we can later modify the HW
profile more freely without breaking VMs.

From a point of view, this is a de-factorization. This is duplicating some
puppet classes, but the previous code sharing situation was only adding more
complexity to maintenance and development.

Bug: T220051
Change-Id: I9f64fdde87cc7be1bf009d0f76687a3d91edcd49
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>