canary host: test method for having canary hosts
This CR adds a new entry to the hiera hierarchy which is search before
the common and role parts.  If the variable canary is not set this
addition will make no difference.  if it set then we create a place
where we can override common and role based parameters.

Hosts: aqs1004.eqiad.wmnet,aqs1005.eqiad.wmnet,puppetmaster1001.eqiad.wmnet,cumin1001.eqiad.wmnet

Bug: T219803
Change-Id: Ifccec7113b58a15c6ceac559f878ff461a3ba7f1