icinga/nagios_common: add Willy Pao to group misleadingly called 'sms'
I say misleading because whether somebody receives SMS or email or
any notification at all is configured at the contact level, in the
private repo with the individual user.

What this means in practice is that this is the "SRE" group that
everybody is a member of and where notifications are sent for
services set to "critical => true".

Which means they are considered important enough to send pages
as opposed to other checks that create notifications on IRC and
in the web UI but don't page anyone.

Most people happen to have a notification method that
includes a phone number but it could also be email-only or a
custom command.

Bug: T221142
Change-Id: I468f5e7b817fe91332be7dc61422fcd6c880167b