db*.yaml: Add hostname as a comment
It has been a few times already that when decommissioning a db* we
forget to delete the YAML file, normally because we tend to grep for the
host name, and the yaml files of course don't show up.
Adding the hostname of each host to their yaml file, so they will show
up on a grep.

This should be also extended to pc*, es* etc...

Change-Id: I496a28aa000f84d8d35bed748b2b4780f23985e8