Add sd-pam processes to filter list for debdeploy
When systemd spawns a user session involving PAM, it forks a child
process inside the session which supervises the session waiting
for the main process to terminate to eventually close the PAM
session. These processes are visible in the process overview as
(sd-pam).

Filter these out, they resolve itself when users complete their
SSH/screen/tmux session. If we ever run into a severe vulnerability
which warrants an immediate refresh of all sessions we'd probably
coordinate reboots anyway.

Note that these appear in lsof as (sd-pam without the closing
bracket, so they also need to be added to the config in this
format.

Bug: T135991
Change-Id: Iecb5a24c7a173dba6cc7f69a0af1a1ff66a62d26