profile::analytics::refinery::repository: use the 'analitics-deploy' user
Once upon a time, we (Analytics) envisioned a future with one
'analytics' user able to perform various kind of recurrent/automated tasks.
The idea was to avoid user profiferation and security management
simplification. Then all the Kerberos work happened, and in T220971
we started working on a solution to replace the powerful 'hdfs'
system user with the 'analytics' one. Everything should work nicely
but having the 'analytics' user managed by scap means that anybody
able to use the analytics_deploy ssh key could also potentially execute
jobs as the analytics team. Keeping separation between Scap and Hadoop
seems to be a good precautionary security measure.

The 'analytics' user is currently not important for deployments, it is
only the owner of the files deployed on various repos. Changing it
to analytics-deploy should be fine (modulo some manual work to chown
files on the scap target hosts).

Bug: T220971
Change-Id: I814109e647f3e5414689163c222e32367a16da8c