Revert "conftool-data: define ats-be for text/upload in all DCs"
This reverts commit 0d4a63cb818321e6491ab03ccdf8e692d873177a.

The assumption behind the commit was that adding the definitions like
that in cache.yaml would create the missing keys (eg: /esams/cache_text/ats-be).

That is not the case, it seems that the keys get created when there is
at least one host defined under them. We need to change the approach and
avoid listing ats-be on DCs/clusters without ATS systems.

Bug: T219967
Change-Id: I1eb0f3103842c9d577f4798865cad24273e771a8