cache: hiera setting to list backend services
Introduce a new hiera setting, defaulting to ['varnish-be'], called
profile::cache::varnish::frontend::backend_services.

Use the setting to choose which etcd keys to use to list available cache
backends for frontends. In clusters/DCs where a mix of ATS and Varnish
backends is available, specify ['ats-be', 'varnish-be']. If all backends
have been converted to ATS, pass 'ats-be' alone.

The goal here is to ensure that confd does not fail in case there is no
host with service 'ats-be' (or later 'varnish-be') in etcd for the given
cluster/DC.

The setting replaces backend_service, which was used to pass 'service'
via the directors. That is not necessary for Varnish backends as they
always only deal with varnish-be.

Bug: T219967
Change-Id: I5b30e415163144c3cfa0d460a482511b26e062e7