sudo: decouple sudo from sudo-ldap
Callers of the sudo module can choose which sudo flavor they want:

* sudo (normal sudo, the default)
* sudo-ldap (a LDAP/nscd/nslcd based sudo)

The sudo-ldap flavor is usually used in CloudVPS, but if using sssd rather
than nscd/nslcd we need sudo instead of sudo-ldap.

This patch involves decoupling the sudo class into another subclass called
sudo::sudoldap, which contains the specific bits related to that version of
sudo.

The desired outcome is that a CloudVPS VM can choose which LDAP client stack to
use between 2:
* classic, using nscd/nslcd, and sudo-ldap
* sssd, using the sssd daemon and sudo (no sudo-ldap)

Both flavors are incompatible.

Bug: T221225
Change-Id: Ibeefa58f3f88cb90cb86e5d78bf43dcd4d16eefb
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>