transparency report: allow members of LDAP 'nda' to see private site
Currently only members of the LDAP groups 'ops' or 'wmf' are allowed
to login on https://transparency-private.wikimedia.org/

In T221744 there is a request to give access to a contractor,
Patrick Johnson of Progresslabs, uid "pbj".

pbj is already a member of 'nda' and in the ldap_only_admins section
of the admin module.

So the nda group as a whole needs to be allowed to login here as well.

Using 'ops / wmf / nda' is a very common default for many logins.

Bug: T221744
Change-Id: I672589da89df837f0c435bb93e4be170336255fe