openstack: clientpackages: decouple configuration
Drop the 'common' class, which is confusing because the word is overloaded and
put the code into proper openstack/debian specific classes.

This means more LOCs and some code duplication, but at this point the decoupled
approach is more interesting that the factorized one, given the complexity to
maintain the dependency/pinning setup for some servers.

Since this profile is only applied to HW servers, this should be a NOOP for
CloudVPS VM instances.

Bug: T220051
Change-Id: I2078ba912c4c59dc81d8fd7d22819abd540f95b8
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>