cloudvps: introduce proper base role/profile for VM instances
This patch introduces a proper base role/profile for VM instances, replacing
the old role::labs::instance class, which doesn't comply with our puppet coding
guidelines.

Every VM instance is expected to use this new role plus any other role
that may be assigned to them via Horizon.

This should help have proper hiera overriding capabilities, which are required
in order to control which flavour of sudo is installed in the VM. This is part
of our work to replace nscd/nslcd with sssd in Cloud VPS, which requires the
sudo package rather than sudo-ldap.
Follow-up patch may add additional hiera logic to this new profile.

Bug: T221225
Change-Id: I0721a2e9f082b9083328d45c651de35343a28bf4
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>