vagrant: refactor roles into profiles
This doesn't follow our puppet coding guidelines and also prevent proper hiera
overriding. Let's create proper profiles.

I would like to get rid of the 'labs' keyword in the role name, but if I
update the role name now, tons of docs would need to be updated as well, and I
don't have time for that now. Leave the role names as they are.

I'm doing this because vagrant sets sudo users/groups, and in a follow-up patch
we will be introducing the option to select different sudo 'flavors' for VM
instances, hoping to be able to replace nscd/nslcd with sssd.

Follow-up patches will introduce hiera lookups into these profiles.

Bug: T221225
Change-Id: I5de786d1866808220893996dd52da52038faad7d
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>